+++ src/hayhooks/redis_scripts/retire_incompatible.lua
@@ -1,9 +1,16 @@
 -- hayhooks:retire_incompatible
 local payload = redis.call('GET', KEYS[1])
-if not payload then return 0 end
+if not payload then
+    redis.call('HDEL', KEYS[6], ARGV[6])
+    redis.call('HDEL', KEYS[7], ARGV[6])
+    return 0
+end
 local record = cjson.decode(payload)
 local status = record['status']
-if status == 'completed' or status == 'failed' or status == 'canceled' then return 0 end
+if status == 'completed' or status == 'failed' or status == 'canceled' then
+    redis.call('HDEL', KEYS[6], ARGV[6])
+    return 0
+end
 if record['definition_revision'] == ARGV[1] or status == 'running' then return 0 end
 
 record['result'] = cjson.null
@@ -47,4 +54,6 @@ redis.call('ZADD', KEYS[4], ARGV[5], record['execution_id'])
 redis.call('HSET', KEYS[5], record['execution_id'], record['status'])
 redis.call('SET', KEYS[1], cjson.encode(record), 'EX', ARGV[3])
 redis.call('ZREM', KEYS[2], record['execution_id'])
+redis.call('HDEL', KEYS[6], record['execution_id'])
+redis.call('HSET', KEYS[7], record['execution_id'], record['sequence'] or 0)
 return 1
