+++ src/hayhooks/server/a2a/__init__.py
@@ -1,7 +1,8 @@
 from hayhooks.a2a import LifecycleAgentExecutor
 from hayhooks.server.a2a.app import create_a2a_app
 from hayhooks.server.a2a.cards import create_agent_card, get_a2a_base_url, is_a2a_exposable
-from hayhooks.server.a2a.executor import ChatCompletionAgentExecutor, DurableAgentExecutor, create_agent_executor
+from hayhooks.server.a2a.durable_executor import DurableAgentExecutor
+from hayhooks.server.a2a.executor import ChatCompletionAgentExecutor, create_agent_executor
 from hayhooks.server.a2a.imports import a2a_import
 from hayhooks.server.a2a.redis_task_store import RedisTaskStore, RedisTaskStoreProvider
 from hayhooks.server.a2a.runtime import (
