+++ src/hayhooks/server/durable/__init__.py
@@ -1,0 +1,1 @@
+"""HTTP-facing durable execution adapters."""
